notebooks/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
ultralytics
